+++ crates/core/src/lib.rs
@@ -59,3 +59,48 @@ pub use chain_data::*;
 
 mod checkpoint;
 pub use checkpoint::*;
+
+/// Core structures for [`TxGraph`].
+///
+/// [`TxGraph`]: https://docs.rs/bdk_chain/latest/bdk_chain/tx_graph/struct.TxGraph.html
+pub mod tx_graph {
+    use crate::collections::{BTreeMap, BTreeSet, HashMap};
+    use alloc::{sync::Arc, vec::Vec};
+    use bitcoin::{OutPoint, Transaction, TxOut, Txid};
+
+    /// Data object used to update a [`TxGraph`].
+    ///
+    /// [`TxGraph`]: https://docs.rs/bdk_chain/latest/bdk_chain/tx_graph/struct.TxGraph.html
+    #[derive(Debug, Clone)]
+    pub struct Update<A = ()> {
+        /// Full transactions.
+        pub txs: Vec<Arc<Transaction>>,
+        /// Floating txouts.
+        pub txouts: BTreeMap<OutPoint, TxOut>,
+        /// Transaction anchors.
+        pub anchors: BTreeSet<(A, Txid)>,
+        /// Seen at times for transactions.
+        pub seen_ats: HashMap<Txid, u64>,
+    }
+
+    impl<A> Default for Update<A> {
+        fn default() -> Self {
+            Self {
+                txs: Default::default(),
+                txouts: Default::default(),
+                anchors: Default::default(),
+                seen_ats: Default::default(),
+            }
+        }
+    }
+
+    impl<A: Ord> Update<A> {
+        /// Extend this update with `other`.
+        pub fn extend(&mut self, other: Update<A>) {
+            self.txs.extend(other.txs);
+            self.txouts.extend(other.txouts);
+            self.anchors.extend(other.anchors);
+            self.seen_ats.extend(other.seen_ats);
+        }
+    }
+}
